Create CSV-Update
Script currently asks for a team, and then opens the csv file that contains the team's roster.  It then loops through the players on the roster and opens up individual player csv files.  

I would like to be able to add new columns and corresponding data (which is generated based on information in other columns) to each of these files.  (ex. new column "FPTs" = 3x"G" + 2x"A" + 0.5x"S", where columns "G", "A" and "S" are existing within the current csv file.